fix: 중복코드 분리
