+++ common_libs/libs/database/conn.py
@@ -1,0 +1,63 @@
+from app.common.config import logger
+
+from fastapi import FastAPI
+from sqlalchemy import create_engine
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker
+
+
+class SQLAlchemy:
+    def __init__(self, app: FastAPI = None, **kwargs):
+        self._engine = None
+        self._session = None
+        if app is not None:
+            self.init_app(app=app, **kwargs)
+
+    def init_app(self, app: FastAPI, **kwargs):
+        database_url = kwargs.get("DB_URL")
+        pool_recycle = kwargs.get("DB_POOL_RECYCLE", 900)
+        is_testing = kwargs.get("TESTING", False)
+        echo = kwargs.get("DB_ECHO", True)
+        is_reload = kwargs.get("RELOAD", False)
+
+        self._engine = create_engine(
+            database_url,
+            echo=echo,
+            pool_recycle=pool_recycle,
+            pool_pre_ping=True,
+        )
+
+        self._session = sessionmaker(autocommit=False, autoflush=False, bind=self._engine)
+
+        @app.on_event("startup")
+        def startup():
+            self._engine.connect()
+            logger.info("DB connected.")
+
+        @app.on_event("shutdown")
+        def shutdown():
+            self._session.close_all()
+            self._engine.dispose()
+            logger.info("DB disconnected.")
+
+    def get_db(self):
+        if self.session is None:
+            raise Exception("must be called 'init_db'")
+        db_session = None
+        try:
+            db_session = self._session()
+            yield db_session
+        finally:
+            db_session.close()
+
+    @property
+    def session(self):
+        return self.get_db
+
+    @property
+    def engine(self):
+        return self._engine
+
+
+db = SQLAlchemy()
+Base = declarative_base()
